+++ xalpha/__init__.py
@@ -3,7 +3,7 @@ __author__ = 'refraction-ray'
 __name__ = 'xalpha'
 
 from xalpha.record import record
-from xalpha.info import fundinfo, indexinfo, cashinfo
+from xalpha.info import fundinfo, indexinfo, cashinfo, mfundinfo
 from xalpha.evaluate import evaluate
 from xalpha.trade import trade
 from xalpha.multiple import mul, mulfix